requirements.txt
Django==1.7.7
Whoosh==2.6.0
django-haystack==2.3.1
pip-tools==0.3.5
wsgiref==0.1.2
